Calculate label numbers for basic blocks
The LLVM syntax allows blocks to be implicitly or explicitly labeled.
In the implicit case the label is the next highest numerical id, given
the ids used so far in the file.

The LLVM parser adds this on the fly.  I decided to determine the
numbers in a post-processing pass.