Remove gen_int(), eval_int() in favor of input()
Previously, in the garbled circuit back ends, you would program (in C)
an input from the generator using

    extern int gen_int()

and there was an equivalent eval_int() for the evaluator.

In the GMW back end, there can be more than 2 parties, so I used

    extern int input(int)

where the argument to input() is the party to read input from.

I've unified the interface by using input() in the garbled circuit case
going forward.  Use input(0) to read from the generator (always party
0), and input(1) to read from the evaluator.