Add -id flag and handle command-line arguments correctly
For the garbled circuit back ends (yao, yaor, gax, gaxr) the compiled
output now expects an -id flag indicating the party (1 for the
evaluator, 0 for the generator, 0 is the default).  The remaining
arguments are the program inputs; they must be integers.

For example, suppose the program is foo.c.  Then after compiling you
need to run the evaluator first (because it is the server) with
something like

    go run foo*.go -id=1 4 5 6

The "-id=1" tells the program to run as the evaluator/server, and 4 5 6
are the evaluator inputs.

Then run

    go run foo*.go -id=0 7 8 9

to run the generator/client with inputs 7 8 9.

Currently the client and server must run on the same machine; host and
port specifications are next.