Use unary representation of active block
Previously we used the variable vStateO to hold the active block,
represented in binary.  At the beginning of each block, if necessary, we
would test this against the (statically known) block number to get a
boolean mask.  This is an expensive way to essentially convert to unary.

Now we uniformly use a unary representation of the active block.  The
difficulty here is that we must sometimes convert (efficiently) from
binary to unary, e.g., when we must compile an LLVM switch instruction.