Move flattening directly into front end
Instead of using cilly to invoke the flattener we call it directly from
smpcc.

This will allow us to more easily transform the initial C code in the
future.

We use ocamlfind to link in the CIL libraries.  This caused some naming
conflicts, so util.ml and gc.ml were renamed (to llabs.ml and
garbled.ml).