LLVM parser: hack to support old syntax
LLVM changed the syntax of metadata to be typeless, however, it retained
support for parsing the types for the moment for compatibility.
Unfortunately the printer is still printing types in some cases, as in
the i64 here:

    !1 = !{!2, !2, i64 0}

So, I've hacked the parser to accept this for the moment.  This
introduces a shift-reduce conflict which does not bode well.