Compiler: fix type in generated call of unary()
The return type of unary() in the call was calculated incorrectly.
It now (correctly) indicates the number of significant bits returned.

The current status is not satisfactory because the return type of
unary() depends on its arguments.  It really should be a slice of
booleans, whose length is determined by its arguments; but that's not
expressible in LLVM.